% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dutch_alpino.r
\name{alpino_clause_rules}
\alias{alpino_clause_rules}
\title{Returns a list with the clause rules for ALPINO}
\usage{
alpino_clause_rules()
}
\value{
A list with rynstax rules, as created with \link{rule}
}
\description{
Returns a list with the clause rules for ALPINO
}
