% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dutch_alpino.r
\name{alpino_quote_queries}
\alias{alpino_quote_queries}
\title{Returns a list with the quote queries for Alpino}
\usage{
alpino_quote_queries(verbs = DUTCH_SAY_VERBS, exclude_verbs = NULL)
}
\arguments{
\item{verbs}{A character vector with verbs used to indicate quotes. A default list of verbs is provided
in DUTCH_SAY_VERBS. If NULL, all verbs are used (except those listed in exclude verbs)}

\item{exclude_verbs}{A character vector with verbs that are exluded. If NULL, no verbs are excluded.}
}
\value{
A list with rynstax queries, as created with \link{tquery}
}
\description{
Returns a list with the quote queries for Alpino
}
