% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dutch_alpino.r
\name{alpino_quote_rules}
\alias{alpino_quote_rules}
\title{Returns a list with the quote rules for ALPINO}
\usage{
alpino_quote_rules()
}
\value{
A list with rynstax rules, as created with \link{rule}
}
\description{
Returns a list with the quote rules for ALPINO
}
