% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_tokens.r
\name{AND}
\alias{AND}
\title{Use AND search in tquery}
\usage{
AND(...)
}
\arguments{
\item{...}{}
}
\value{
A list, to be used as input to \link{tquery}
}
\description{
Use AND search in tquery
}
\examples{
tquery(AND(lemma = 'walk', POS='Noun'))   ## is also the default
}
