% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.r
\name{annotate}
\alias{annotate}
\title{Annotate a tokenlist with the results of \link{find_nodes}}
\usage{
annotate(tokens, nodes, column, use = NULL, fill = T, check = T,
  block = NULL)
}
\arguments{
\item{tokens}{A tokenIndex data.table, created with \link{as_tokenindex}, or any data.frame with the required columns (see \link{tokenindex_columns}).}

\item{nodes}{A data.table, as created with \link{find_nodes} or \link{apply_rules}. Can be a list of multiple data.tables.}

\item{column}{The name of the column in which the annotations are added. The unique ids are added as [column]_id}

\item{use}{Optionally, specify which columns from nodes to add. Other than convenient, this is slighly different 
from subsetting the columns in 'nodes' beforehand if fill is TRUE. When the children are collected,
the ids from the not-used columns are still blocked (see 'block')}

\item{fill}{If TRUE, the children for each id are recursively (children of children etc.) added. If this leads to
duplicate ids (if an id in nodes is a child of another id in nodes), the most direct children are kept.
For example, if 1 -> 2 -> 3, and both 1 and 2 are in 'nodes', then 3 is only added as a child of 2.}

\item{check}{If TRUE, give a warning if there are duplicates in the data (in which case duplicates are deleted)}

\item{block}{Optionally, another set of nodes, of which the .KEY values will be blocked for annotations}
}
\description{
Use the results of \link{find_nodes} (nodes) to add two columns to a tokenlist.
One column contains the ids for each match.
The other column contains the annotations, using the column names of 'nodes'.
}
