% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dutch_alpino.r
\name{annotate_alpino}
\alias{annotate_alpino}
\title{Quote and clause extraction for Dutch texts, parsed with ALPINO}
\usage{
annotate_alpino(tokens)
}
\arguments{
\item{tokens}{Either a tokenIndex data.table, created with \link{as_tokenindex}, or any data.frame with the required columns (see \link{tokenindex_columns}).}

\item{no_overlap}{If TRUE (default), then quote patterns (e.g., SOURCE says QUOTE) are excluded for clauses. Clauses can then still occur within quotes, such as in: SOURCE says: "SUBJECT is happy". 
If a quote occurs within a clause, the clause will stop, as in: (SUBJECT is mad at) (SOURCE, who said QUOTE)}
}
\value{
the tokens data as a tokenIndex, with additional columns: quote, quote_id, clause and clause_id
}
\description{
This is a convenience function that uses all the (currently implemented) rules for quote and clause extraction.
}
