% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{annotate_tokens}
\alias{annotate_tokens}
\title{Annotate a tokenlist with quotes and clauses}
\usage{
annotate_tokens(tokens, quotes, clauses)
}
\arguments{
\item{tokens}{a df of tokens}

\item{quotes}{the output of get_quotes}

\item{clauses}{the output of get_clauses}
}
\value{
a df of tokens with columns for clause_id and clause_role
}
\description{
Annotate a tokenlist with quotes and clauses
}

