% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_index.r
\name{as_tokenindex}
\alias{as_tokenindex}
\title{Prepare a tokenIndex}
\usage{
as_tokenindex(tokens, doc_id = "doc_id", sentence = "sentence",
  token_id = "token_id", parent = "parent")
}
\arguments{
\item{tokens}{A data.frame, data.table, or tokenindex.}

\item{doc_id}{The name of the document id columns}

\item{sentence}{The name of the sentence (id/index) column}

\item{token_id}{The name of the token id column}

\item{parent}{The name of the parent id column}
}
\description{
Creates a tokenIndex data.table, that is required to use \link{find_nodes}. 
Accepts any data.frame given that the required columns (doc_id, sentence, token_id, parent) are present.
Alternative column names can be given.
}
\details{
The data will be sorted by the doc_id, sentence and token_id columns.
Accordingly, it is recommended to use numeric token_id's. 
Some parsers return token_id's as numbers with a prefix (t_1, w_1), in which case sorting is inconvenient (t_15 > t_100).

The data in the data.frame will not be changed, with two exceptions. First, the columnnames will be changed if the default values are not used.
Second, if a token has itself as its parent (which in some parsers is used to indicate the root), the parent is set to NA (as used in other parsers) to prevent infinite cycles.
}
