% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nodes.r
\name{block_ids}
\alias{block_ids}
\title{Get and/or merge ids for the block argument in \link{find_nodes}}
\usage{
block_ids(..., names = NULL)
}
\arguments{
\item{...}{Either a data.table with the columns doc_id and token_id, or the output of \link{find_nodes}}
}
\value{
A data.table with the columns doc_id and token_id
}
\description{
Get and/or merge ids for the block argument in \link{find_nodes}
}
