% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.r
\name{cast_tokens_dfm}
\alias{cast_tokens_dfm}
\title{Like cast_tokens, but instead returns the tokens as a quanteda document-feature matrix}
\usage{
cast_tokens_dfm(tokens, x, FUN, by, id, collapse_id = F, ...)
}
\arguments{
\item{tokens}{A tokenIndex with rsyntax annotations}

\item{x}{The name of the column that will be aggregated (!not the vector itself)}

\item{FUN}{The function used to aggregate.}

\item{by}{A list or named vector to specify the grouping elements, where names are columns and values are values in the columns. For example, if there are 
"quotes" and "clauses" columns, that have "source" and "subject" respectively, use c(quotes = "source", clauses = "subject").}

\item{id}{The name of the column the contains the ids/codes of the groups. By default, each unique code is casted to a subsentence. Alternatively,
if the id column contains text instead of unique codes, the collapse_id argument can be used to collapse the text into an id.}

\item{collapse_id}{see id}

\item{...}{Additional arguments passed to FUN}
}
\description{
Like cast_tokens, but instead returns the tokens as a quanteda document-feature matrix
}
