% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.r
\name{cast_tokens_text}
\alias{cast_tokens_text}
\title{Like cast_tokens, but prints for easier inspection}
\usage{
cast_tokens_text(tokens, text, by, id, collapse_id = F, rm.na = F)
}
\arguments{
\item{tokens}{A tokenIndex with rsyntax annotations}

\item{text}{THe name of the column with token texts (often token or lemma)}

\item{by}{A list or named vector to specify the grouping elements, where names are columns and values are values in the columns. For example, if there are 
"quotes" and "clauses" columns, that have "source" and "subject" respectively, use c(quotes = "source", clauses = "subject").}

\item{id}{The name of the column the contains the ids/codes of the groups. By default, each unique code is casted to a subsentence. Alternatively,
if the id column contains text instead of unique codes, the collapse_id argument can be used to collapse the text into an id.}

\item{collapse_id}{see id}
}
\value{
a data.table
}
\description{
Like cast_tokens, but prints for easier inspection
}
