% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applied_reshape.r
\name{chop}
\alias{chop}
\title{Chop of a branch of the tree}
\usage{
chop(.tokens, ...)
}
\arguments{
\item{.tokens}{}

\item{...}{Arguments passed to tquery. For instance, relation = 'punct' cuts off all punctuation dependencies (in universal dependencies)}
}
\description{
Using the query language for tquery, chop of the branch down from the node that is found
}
