% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_index.r
\name{cname}
\alias{cname}
\title{Get a tokenindex column name}
\usage{
cname(...)
}
\arguments{
\item{...}{character strings}
}
\value{
a character vector
}
\description{
gets a tokenindex colunm name, taking into account aliases specified with tokenindex_column. Uses the input if no alias is specified.
}
\examples{
cname('doc_id')
}
