% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clauses_alpino.R
\name{collapse_conjunctions}
\alias{collapse_conjunctions}
\title{Collapse conjunctions}
\usage{
collapse_conjunctions(tokens, nodes_df, id_col, remove_conj_node = T)
}
\arguments{
\item{tokens}{a token list data frame}

\item{nodes_df}{the output of get_clauses}

\item{id_col}{the name of the column in which conjunctions are looked up}

\item{remove_conj_node}{if TRUE, the row with the conjunction word will be removed from the clauses}
}
\description{
Collapse conjunctions
}

