% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{construct_triples}
\alias{construct_triples}
\title{Construct a list of source/subject/object triples from a tokens list with clauses and quotes}
\usage{
construct_triples(tokens, concept_column = "concept")
}
\arguments{
\item{tokens}{a data frame with clause_id, clause_role, quote_id and quote_role columns}

\item{concept_column}{the name of the column in tokens that contains identified 'concepts'}
}
\value{
a data frame with clause_id and source, subject, object columns indicating concepts found in those positions
}
\description{
Construct a list of source/subject/object triples from a tokens list with clauses and quotes
}

