% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_english_corenlp.r
\name{corenlp_clause_queries}
\alias{corenlp_clause_queries}
\title{Returns a list with the clause queries for CoreNLP}
\usage{
corenlp_clause_queries(verbs = NULL, exclude_verbs = ENGLISH_SAY_VERBS,
  with_subject = T, with_object = F)
}
\arguments{
\item{verbs}{A character vector with verbs used to indicate clauses. If NULL (default), all verbs are used (except those listed in exclude verbs)}

\item{exclude_verbs}{A character vector with verbs that are not used in clauses. By default, this is the list of ENGLISH_SAY_VERBS, 
which are the verbs used in the corenlp_quote_queries(). If set to NULL, no verbs are excluded.}

\item{tokens}{a token list data frame}
}
\value{
a data.table with nodes (as .G_ID) for id, subject and predicate
}
\description{
Returns a list with the clause queries for CoreNLP
}
