% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenlp.r
\name{corenlp_tokenindex}
\alias{corenlp_tokenindex}
\title{Parse text using udpipe}
\usage{
corenlp_tokenindex(text, replace_coref = F)
}
\arguments{
\item{text}{A character vector}

\item{replace_coref}{If TRUE, replace coreference with main entity}
}
\value{
A tokenindex
}
\description{
Uses the cleanNLP package to parse text with the coreNLP parser.
To use, first download corenlp with \code{\link[cleanNLP]{cnlp_download_corenlp}}  
and initialize corenlp with \code{\link[cleanNLP]{cnlp_init_corenlp}}
}
\details{
If you want to use coreference resolution, make sure to initialize corenlp with anno_level = 3.
}
\examples{
\donttest{
cnlp_download_corenlp()                ## first time only

cnlp_init_corenlp('en', anno_level=3)  ## initialize before use

corenlp_tokenindex('Bob said that he was tired')
}
}
