% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_search.r
\name{deep_family}
\alias{deep_family}
\title{Get the parents or children of a set of ids}
\usage{
deep_family(tokens, id, level, depth, minimal = F, block = NULL,
  replace = F, show_level = F, only_new = F, lookup = NULL,
  g_id = NULL)
}
\arguments{
\item{tokens}{The tokenIndex}

\item{id}{rows in the tokenIndex to use as the ID (for whom to get the family)}

\item{level}{either 'children' or 'parents'}

\item{depth}{How deep to search. eg. children -> grandchildren -> grandgrand etc.}

\item{minimal}{If TRUE, only return doc_id, sentence, token_id and parent}

\item{block}{A data.table with global ids (doc_id,sentence,token_id) for excluding nodes from the search}

\item{replace}{If TRUE, re-use nodes}

\item{show_level}{If TRUE, return a column with the level at which the node was found (e.g., as a parent, grantparent, etc.)}

\item{only_new}{If TRUE, only return new found family. Otherwise, the id input is included as well.}

\item{lookup}{Optional lookup filter}

\item{g_id}{Optional filter with specific global token ids}
}
\description{
Get the parents or children of a set of ids
}
