% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{example_tokens}
\alias{example_tokens}
\title{Tokens from "This sentence is parsed by Stanford CoreNLP"}
\usage{
data(example_tokens)
}
\description{
Tokens from "This sentence is parsed by Stanford CoreNLP"
}
\keyword{datasets}

