% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_tokens_corenlp}
\alias{example_tokens_corenlp}
\title{Tokens for CorenLP sentences containing simple quotes}
\usage{
data(example_tokens_corenlp)
}
\description{
Tokens for CorenLP sentences containing simple quotes
}
\keyword{datasets}

