% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_tokens_dutchclauses}
\alias{example_tokens_dutchclauses}
\title{Tokens from Dutch sentences for testing get_clauses}
\usage{
data(dutchclauses)
}
\description{
Tokens from Dutch sentences for testing get_clauses
}
\keyword{datasets}

