% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_tokens_dutchquotes}
\alias{example_tokens_dutchquotes}
\title{Tokens from Dutch sentences containing quotes}
\usage{
data(dutchquotes)
}
\description{
Tokens from Dutch sentences containing quotes
}
\keyword{datasets}

