% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{fill}
\alias{fill}
\title{Fill out the ids to all their descendants, until they hit one of the ids in the block list}
\usage{
fill(id, tokens, block = unique(id))
}
\arguments{
\item{id}{a vector of ids to fill out}

\item{tokens}{a data frame of tokens containing child and id columns}
}
\value{
a data frame with the original ids and all descendants
}
\description{
Fill out the ids to all their descendants, until they hit one of the ids in the block list
}

