% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{find_nodes}
\alias{find_nodes}
\title{Search for nodes matching specific criteria}
\usage{
find_nodes(tokens, children = NULL, columns = NULL, ...)
}
\arguments{
\item{tokens}{a df of tokens}

\item{children}{a list of children, each a list of child attribute filters}

\item{columns}{vector of column names to return from the node}

\item{...}{node attribute filters}
}
\value{
a df with (parent) id and a column for each rel with the respective child id
}
\description{
You can search on node attributes (lemma, pos) and on attributes of children.
The function will return a data frame with the found nodes and a column for each matched child relation.
The node attributes allow __i (e.g. lemma_i="bush") to match case insensitive, 
and __in (e.g. id__in=ids) to match multiple values.
The child attributes can be either a single relation name, or a list of node attributes.
The return column for children is named after the explicit rename= attribute, defaulting to the relation pattern, if given.
}

