% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolate_branches.r
\name{get_branch_id}
\alias{get_branch_id}
\title{Add the branch id as a column to the tokenindex}
\usage{
get_branch_id(tokens)
}
\arguments{
\item{tokens}{A tokenindex}
}
\value{
the tokenindex
}
\description{
After splitting trees into branches
}
