% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clauses_alpino.R
\name{get_clauses_nl}
\alias{get_clauses_nl}
\title{Get clauses from tokens parsed by Alpino}
\usage{
get_clauses_nl(tokens, quotes = NULL)
}
\arguments{
\item{tokens}{a token list data frame}

\item{quotes}{the return value of get_quotes_nl}
}
\value{
a data frame with columns clause_id, source, predicate
}
\description{
Get clauses from tokens parsed by Alpino
}

