% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.r
\name{get_nodes}
\alias{get_nodes}
\title{Transform the nodes to long format and match with token data}
\usage{
get_nodes(tokens, nodes, use = NULL, fill = T, token_cols = c("token"),
  block = NULL, show_fill = F)
}
\arguments{
\item{tokens}{A tokenIndex data.table, created with \link{as_tokenindex}, or any data.frame with the required columns (see \link{tokenindex_columns}).}

\item{nodes}{A data.table, as created with \link{find_nodes} or \link{apply_rules}. Can be a list of multiple data.tables.}

\item{use}{Optionally, specify which columns from nodes to add. Other than convenient, this is slighly different 
from subsetting the columns in 'nodes' beforehand if fill is TRUE. When the children are collected,
the ids from the not-used columns are still blocked (see 'block')}

\item{fill}{If TRUE, the children for each id are recursively (children of children etc.) added. If this leads to
duplicate ids (if an id in nodes is a child of another id in nodes), the most direct children are kept.
For example, if 1 -> 2 -> 3, and both 1 and 2 are in 'nodes', then 3 is only added as a child of 2.}

\item{token_cols}{A character vector, specifying which columns from tokens to include in the output}

\item{block}{Optionally, another set of nodes, of which the .KEY values will be blocked for annotations}

\item{show_fill}{If TRUE, add a .FILL column that indicates which tokens were added with fill}
}
\value{
A data.table with the nodes in long format, and the specified token_cols attached
}
\description{
Transform the nodes to long format and match with token data
}
