% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_parents}
\alias{get_parents}
\title{get all parents with the given relation,}
\usage{
get_parents(tokens, relation = NULL, rename = relation, ...)
}
\arguments{
\item{tokens}{a df of tokens}

\item{rename}{optionally rename the output column, defaults to re}

\item{...}{other filters}

\item{rel}{the relation of child with parent (optional)}
}
\value{
a df with the ids of (parent) id and child (id)
}
\description{
get all parents with the given relation,
}

