% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clauses_alpino.R
\name{get_quotes2}
\alias{get_quotes2}
\title{Get all quotes in the tokens}
\usage{
get_quotes2(tokens)
}
\arguments{
\item{tokens}{a data frame of tokens containing id, parent, relation, pos1 and lemma columns.}
}
\value{
the quotes in long format, i.e. a data frame with quote id, quote role, and token id
}
\description{
Runs the syntax rules to get quotes/paraphrases from the sentences.
Note: The token ids should be globally unique, see unique_ids!
}

