% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dutch_alpino.r
\name{get_quotes_alpino}
\alias{get_quotes_alpino}
\title{Get quotes from tokens parsed by Alpino}
\usage{
get_quotes_alpino(tokens, block = NULL)
}
\arguments{
\item{tokens}{A tokenIndex data.table, created with \link{as_tokenindex}, or any data.frame with the required columns (see \link{tokenindex_columns}).}

\item{block}{Another set of nodes (as created with \link{find_nodes}), of which the .KEY values will be blocked for annotations}
}
\value{
a data.table with nodes
}
\description{
Get quotes from tokens parsed by Alpino
}
