% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotes_alpino.R
\name{get_quotes_nl}
\alias{get_quotes_nl}
\title{Get quotes from tokens parsed by Alpino}
\usage{
get_quotes_nl(tokens)
}
\arguments{
\item{tokens}{a token list data frame}
}
\value{
a data frame with columns id, source, and quote
}
\description{
Get quotes from tokens parsed by Alpino
}

