% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rsyntax.R
\name{get_text}
\alias{get_text}
\title{Get the text from a tokens data frame}
\usage{
get_text(tokens, word.column = "word")
}
\arguments{
\item{tokens}{a data frame of tokens containing id, parent, and relation columns}

\item{word.column}{the column(s) containing the words to use.}
}
\value{
the text as a single character value
}
\description{
The data is sorted by aid, sentence, and offset, and the words are pasted together to a single character value.
The word.column parameter specifies which column to use for the words.
If multiple columns arse given (e.g. lemma and POS) they are pasted together with a slash ("Like/IN this/DT")
}

