% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rsyntax.R
\name{graph_from_sentence}
\alias{graph_from_sentence}
\title{Create an igraph tree from a sentence}
\usage{
graph_from_sentence(tokens, sentence = NULL)
}
\arguments{
\item{tokens}{a data frame of tokens containing id, parent, and relation columns}

\item{sentence}{an optional sentence to filter the tokens (which should then contain a sentence column)}
}
\value{
an igraph graph
}
\description{
Create an igraph tree from a data frame with token information, e.g, as donwloaded with amcat.get.tokens.
This assumes that id, parent, and relation are presente in the data frame.
Any other columns will be available as vertex and edge attributes (edge attributes are taken from the child token).
If available, the 'word' column will be used to label the vertices.
The data frame should contain a single sentence with a single root node (with no parent).
}

