% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applied_reshape.r
\name{inherit}
\alias{inherit}
\title{Have nodes with a certain relation inherit the position of their parent}
\usage{
inherit(
  .tokens,
  relation,
  take_fill = fill(),
  give_fill = fill(),
  unpack = T,
  only_new = "relation"
)
}
\arguments{
\item{.tokens}{A tokenIndex}

\item{relation}{A character string specifying the relation}

\item{take_fill}{If TRUE, give the node that will inherit the parent position a copy of the parent children (but only if it does not already have children with this relation; see only_new)}

\item{give_fill}{If TRUE, copy the children of the node that will inherit the parent position to the parent (but only if it does not already have children with this relation; see only_new)}

\item{unpack}{If TRUE, create separate branches for the parent and the node that inherits the parent position}

\item{only_new}{A character vector giving one or multiple column names that need to be unique for take_fill and give_fill}
}
\description{
This is in particular usefull for removing conjunctions
}
