% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_help.r
\name{inspect_family}
\alias{inspect_family}
\title{Given a query, find all possible parents/children for a given node}
\usage{
inspect_family(tokens, query, node, family = c("family", "children",
  "parents"), depth = 1, group_cols = NULL)
}
\arguments{
\item{tokens}{A tokenIndex}

\item{query}{An rsyntax tquery}

\item{node}{The name of a node (specified with 'save' in the tquery)}

\item{family}{Either 'children', 'parents',  or 'family' (both children and parents)}

\item{depth}{An integer for how deep to search. 1 is only children, 2 includes grandchilderen, 3 includes grandgrandchildren, etc.}

\item{group_cols}{Optionally, names of columns in tokens to group the results by. Groups are aggregated, and the number of times (and percentage) of each group is given.}
}
\description{
Given a query, find all possible parents/children for a given node
}
