% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolate_branches.r
\name{isolate_branch}
\alias{isolate_branch}
\title{Isolate a branch in a dependency tree}
\usage{
isolate_branch(tokens, ..., copy_parent = T)
}
\arguments{
\item{tokens}{}

\item{...}{lookup arguments to find the node to split. For example, isolate_branch(tokens, relation='relcl') 
isolates branches of which the top node (the new root) has the relation "relcl".}

\item{copy_parent}{If TRUE (default) copy the parent of the branch and include it in the isolated branch}
}
\description{
cuts of a branch at the nodes that match the lookup arguents (...).
A "branch_parent" column is added to the tokenindex, that indicates for the new roots
which node the parent was.
}
