% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_reshape.r
\name{mutate_nodes}
\alias{mutate_nodes}
\title{Mutate nodes}
\usage{
mutate_nodes(.tokens, node, ..., subset = NULL)
}
\arguments{
\item{.tokens}{A tokenIndex in which nodes are selected with \link{select_nodes}.}

\item{node}{The name of the node that is to be mutated}

\item{...}{named arguments. The name should be a column in tokens}

\item{subset}{A subset expression (that evaluates to a logical vector). The token column for each labeled node in the tquery can be referred to as label$column.}
}
\description{
Mutate nodes
}
