% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_tokens.r
\name{NOT}
\alias{NOT}
\title{Use NOT search in tquery}
\usage{
NOT(...)
}
\arguments{
\item{...}{}
}
\value{
A list, to be used as input to \link{tquery}
}
\description{
Use NOT search in tquery
}
\examples{
tquery(AND(lemma = 'walk', POS='Noun'))   ## is also the default
}
