% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax_rules.r
\name{print.rsyntaxRule}
\alias{print.rsyntaxRule}
\title{S3 print for rsyntaxRule class}
\usage{
\method{print}{rsyntaxRule}(x, ...)
}
\arguments{
\item{x}{an rsyntaxRule function}

\item{...}{not used}
}
\description{
S3 print for rsyntaxRule class
}
\examples{
rule = rule(select = lemma \%in\% .VIND_VERBS, 
                   children(save = 'source', p_rel=.SUBJECT_REL),
                   children(p_rel='vc', select = POS \%in\% c('C', 'comp'),
                            children(save='quote', p_rel=.SUBJECT_BODY)))
rule 
}
