% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.r
\name{print.tReshape}
\alias{print.tReshape}
\title{S3 print for tReshape class}
\usage{
\method{print}{tReshape}(x, ...)
}
\arguments{
\item{x}{a tReshape}

\item{...}{not used}
}
\description{
S3 print for tReshape class
}
\examples{
r = treshape(bypass = 'conj', remove='cc', link_children='nsubj')
r
}
