% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{quote_punctuation}
\alias{quote_punctuation}
\title{Quote punctuation}
\format{
character()
}
\usage{
data(quote_punctuation)
}
\description{
Punctuation used in quotes, for use in syntax queries
}
\keyword{datasets}
