% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nodes.r
\name{rec_find}
\alias{rec_find}
\title{Recursive search tokens}
\usage{
rec_find(tokens, ids, ql, block = NULL)
}
\arguments{
\item{tokens}{The tokenIndex}

\item{ids}{A data.table with global ids (doc_id,sentence,token_id).}

\item{ql}{a list of queriers (possibly the list nested in another query, or containing nested queries)}

\item{block}{A data.table with global ids (doc_id,sentence,token_id) for excluding nodes from the search}
}
\description{
Recursive search tokens
}
