% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_reshape.r
\name{remove_fill}
\alias{remove_fill}
\title{Remove fill}
\usage{
remove_fill(.tokens, node, rm_subset_fill = NULL, rm_subset = NULL,
  keep_shared = F)
}
\arguments{
\item{.tokens}{A tokenIndex in which nodes are selected with \link{select_nodes}.}

\item{node}{The name of the node that is to be mutated}

\item{rm_subset_fill}{A subset on the fill nodes. Can only directly use token column. For example, use pos == 'VERB' to remove only verbs}

\item{rm_subset}{A subset expression (that evaluates to a logical vector) to more specifically specify which nodes to remove. The token column for each labeled node in the tquery can be referred to as label$column.}

\item{keep_shared}{If there is another node that has the same fill nodes, should the fill nodes that are shared also be removed?}
}
\description{
Like remove_nodes, but only removing the fill nodes
}
