% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_reshape.r
\name{reselect_nodes}
\alias{reselect_nodes}
\title{Within a chain of reshape operations, reapply the tquery}
\usage{
reselect_nodes(.tokens)
}
\arguments{
\item{.tokens}{A tokenIndex in which nodes are selected with \link{select_nodes}.}
}
\description{
Within a chain of reshape operations, reapply the tquery
}
