% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.r
\name{reshape_tokens}
\alias{reshape_tokens}
\title{Reshape tokens based on rsyntax annotations}
\usage{
reshape_tokens(tokens, by)
}
\arguments{
\item{tokens}{}

\item{by}{}
}
\description{
Reshape tokens based on rsyntax annotations
}
