\name{rsyntax-package}
\alias{rsyntax-package}
\alias{rsyntax}
\docType{package}
\title{
\packageTitle{rsyntax}
}
\description{
\packageDescription{rsyntax}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rsyntax}
\packageIndices{rsyntax}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rsyntax}

Maintainer: \packageMaintainer{rsyntax}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
