% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rsyntax_threads}
\alias{rsyntax_threads}
\title{Get the number of threads to be used by rsyntax functions}
\usage{
rsyntax_threads()
}
\value{
the setting for the number of threads used by rsyntax
}
\description{
rsyntax relies heavily on the data.table package, which supports multithreading. 
By default, the number of threads set by data.table are used, as you can see with \code{\link[data.table]{getDTthreads}}.
With \code{\link{set_rsyntax_threads}} you can set the number of threads for rsyntax functions, without affecting the data.table settings.
}
\examples{
rsyntax_threads()
}
