% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{searchQuery}
\alias{searchQuery}
\title{Search for tokens in a tokenlist using indicators with conditions}
\usage{
searchQuery(tokens, indicator, condition = "", default.window = 25,
  meta = NULL, condition_once = FALSE, presorted = T, filter = rep(T,
  nrow(tokens)), aid_var = "aid", term_i_var = "id", term_var = "word")
}
\arguments{
\item{tokens}{}

\item{indicator}{}

\item{condition}{}

\item{meta}{}

\item{condition_once}{logical. If TRUE, then if an indicator satisfies its conditions once in an article, all indicators within that article are coded.}

\item{presorted}{The data has to be sorted on order(aid_var, term_i_var). If this is already the case, presorted can be set to TRUE to save time (which is usefull when testing many individual queries for large tokenlists)}

\item{filter}{}

\item{aid_var}{}

\item{term_i_var}{}

\item{term_var}{}

\item{sample_n}{}

\item{keywordIC}{}

\item{kwic_nwords}{}
}
\value{
the tokens that match the query
}
\description{
Tokens need to be sorted on order(aid_var, term_i_var), or presorted needs to be set to FALSE. Note that if this function is used often, its faster to sort tokens first and use presorted = TRUE.
}

