% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_reshape.r
\name{select_nodes}
\alias{select_nodes}
\title{Apply tquery to initiate reshape operations}
\usage{
select_nodes(tokens, tquery, fill = T, fill_only_first = T)
}
\arguments{
\item{tokens}{A tokenIndex data.table, or any data.frame coercible with \link{as_tokenindex}.}

\item{tquery}{A \link{tquery} that selects and labels the nodes that are used in the reshape operations}

\item{fill}{Logical, should fill be used?}

\item{fill_only_first}{Logical, should a node only be filled once, with the nearest (first) labeled node?}
}
\description{
Apply tquery to initiate reshape operations
}
