% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nodes.r
\name{select_tokens}
\alias{select_tokens}
\title{Select which tokens to add}
\usage{
select_tokens(tokens, ids, q, e, block = NULL)
}
\arguments{
\item{tokens}{The tokenIndex}

\item{ids}{A data.table with global ids (doc_id,sentence,token_id).}

\item{q}{a query (possibly nested in another query, or containing nested queries)}

\item{e}{an environment (in case query contains $select)}

\item{block}{A data.table with global ids (doc_id,sentence,token_id) for excluding nodes from the search}
}
\description{
Given ids, look for their parents/children (specified in query q) and filter on criteria (specified in query q)
}
