% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_english_spacy.R
\name{spacy_english_clause_queries}
\alias{spacy_english_clause_queries}
\title{Returns a list with the clause queries for spacy_english_}
\usage{
spacy_english_clause_queries(verbs = NULL,
  exclude_verbs = ENGLISH_SAY_VERBS, with_subject = T,
  with_object = F, sub_req = T, ob_req = F)
}
\arguments{
\item{verbs}{A character vector with verbs used to indicate clauses. If NULL (default), all verbs are used (except those listed in exclude verbs)}

\item{exclude_verbs}{A character vector with verbs that are not used in clauses. By default, this is the list of ENGLISH_SAY_VERBS, 
which are the verbs used in the spacy_english__quote_queries(). If set to NULL, no verbs are excluded.}

\item{with_subject}{If True, subject nodes will be "saved" as "subject"}

\item{with_object}{Same for object nodes}

\item{sub_req}{If True, subject nodes are required (must be matched). By default this is true.}

\item{ob_req}{Same for object nodes. By default this is false.}
}
\value{
a data.table with nodes (as .G_ID) for id, subject and predicate
}
\description{
Returns a list with the clause queries for spacy_english_
}
