% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applied_reshape.r
\name{split_UD_conj}
\alias{split_UD_conj}
\title{Split conjunctions for dependency trees in Universal Dependencies}
\usage{
split_UD_conj(
  tokens,
  conj_rel = "conj",
  cc_rel = "cc",
  unpack = T,
  no_fill = NULL,
  min_dist = 0,
  max_dist = Inf,
  compound_rel = c("compound*", "flat")
)
}
\arguments{
\item{tokens}{a tokenIndex based on texts parsed with \code{\link[spacyr]{spacy_parse}} (with dependency=TRUE)}

\item{conj_rel}{The dependency relation for conjunctions. By default conj}

\item{cc_rel}{The dependency relation for the coordinating conjunction. By default cc. This will be removed.}

\item{unpack}{}

\item{no_fill}{}

\item{min_dist}{}

\item{max_dist}{}

\item{compound}{}
}
\value{

}
\description{
Split conjunctions for dependency trees in Universal Dependencies
}
