% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{testQuery}
\alias{testQuery}
\title{Search for tokens in a tokenlist using indicators with conditions}
\usage{
testQuery(tokens, indicator, condition = "", default.window = 25,
  tokenfreq = T, keywordIC = T, kwic_nwords = 10, keywordIC_sample = 10,
  random_sample = T, meta = NULL, condition_once = FALSE, presorted = T,
  aid_var = "aid", term_i_var = "id", term_var = "word")
}
\arguments{
\item{tokens}{}

\item{indicator}{}

\item{condition}{}

\item{default.window}{}

\item{tokenfreq}{}

\item{keywordIC}{}

\item{kwic_nwords}{}

\item{keywordIC_sample}{}

\item{random_sample}{}

\item{meta}{}

\item{condition_once}{}

\item{presorted}{}

\item{aid_var}{}

\item{term_i_var}{}

\item{term_var}{}
}
\value{
prints stuff to evaluate the validity of queries
}
\description{
Tokens need to be sorted on order(aid_var, term_i_var), or presorted needs to be set to FALSE. Note that if this function is used often, its faster to sort tokens first and use presorted = TRUE.
}

