% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nodes.r
\name{token_family}
\alias{token_family}
\title{Get the parents or children of a set of ids}
\usage{
token_family(tokens, ids, level = "children", depth = Inf,
  minimal = F, block = NULL, replace = F, show_level = F)
}
\arguments{
\item{tokens}{The tokenIndex}

\item{ids}{A data.table with global ids (doc_id,sentence,token_id).}

\item{level}{either 'children' or 'parents'}

\item{depth}{How deep to search. eg. children -> grandchildren -> grandgrand etc.}

\item{minimal}{If TRUE, only return doc_id, sentence, token_id and parent}

\item{block}{A data.table with global ids (doc_id,sentence,token_id) for excluding nodes from the search}

\item{replace}{If TRUE, re-use nodes in deep_family()}

\item{show_level}{}
}
\description{
Get the parents or children of a set of ids
}
