% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{tokenClauseAnnotation}
\alias{tokenClauseAnnotation}
\title{Annotate a tokenlist with clauses}
\usage{
tokenClauseAnnotation(tokens, clauses)
}
\arguments{
\item{tokens}{a df of tokens}

\item{clauses}{the output of get_clauses}
}
\value{
a df of tokens with columns for clause_id and clause_role
}
\description{
Annotate a tokenlist with clauses
}

