% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcorpus.r
\name{tokenindex_tcorpus}
\alias{tokenindex_tcorpus}
\title{Create a tcorpus from a tokenindex}
\usage{
tokenindex_tcorpus(tokens, meta = NULL, copy_ids = F)
}
\arguments{
\item{tokens}{A tokenindex (\code{\link{as_tokenindex}})}

\item{meta}{A data.frame with document meta data, where each row is a document, and columns contain the meta data per document. 
Needs to have a column named "doc_id", of which the values correspond to the doc_id in tokens.}

\item{copy_ids}{If TRUE, the original token_id and parent column will be copied to the columns orig_token_id and orig_parent. (see description for why)}
}
\description{
This function creates a \code{\link[corpustools]{tCorpus}}. This is a corpus class for a tokenlist (and document meta data) that provides various functions for querying, analyzing
and visualizing text. 

In the tCorpus, token ids need to be locally unique within documents.
If this is not the case, the 'token_id' and 'parent' columns will be transformed. 
The new token_id and parent can still be used as normally within rsyntax, but there can be reasons to want to preserve the original token_id and parent.
Most importantly, the data might contain other columns that represent edges between tokens (e.g., coreference resolution). 
If copy_ids is TRUE, the original token_id and parent values will therefore be copied to the new columns orig_token_id and orig_parent.
}
