% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tokens_corenlp}
\alias{tokens_corenlp}
\title{Example tokens for coreNLP English}
\format{
data.frame
}
\usage{
data(tokens_corenlp)
}
\description{
Example tokens for coreNLP English
}
\keyword{datasets}
