% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tokens_dutchclauses}
\alias{tokens_dutchclauses}
\title{Example tokens for Dutch clauses}
\format{
data.frame
}
\usage{
data(tokens_dutchclauses)
}
\description{
Example tokens for Dutch clauses
}
\keyword{datasets}
