% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tokens_dutchquotes}
\alias{tokens_dutchquotes}
\title{Example tokens for Dutch quotes}
\format{
data.frame
}
\usage{
data(tokens_dutchquotes)
}
\description{
Example tokens for Dutch quotes
}
\keyword{datasets}
