% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{tokens_from_coreNLP}
\alias{tokens_from_coreNLP}
\title{Create a tokens dataframe from a CoreNLP annotation object}
\usage{
tokens_from_coreNLP(a)
}
\arguments{
\item{a}{coreNLP annotation as returned by coreNLP::annotate*}
}
\value{
a data frame compatible with getclauses / getquotes
}
\description{
Create a tokens dataframe from a CoreNLP annotation object
}
