% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tokens_spacy}
\alias{tokens_spacy}
\title{Example tokens for spacy English}
\format{
data.frame
}
\usage{
data(tokens_spacy)
}
\description{
Example tokens for spacy English
}
\keyword{datasets}
