% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_text.r
\name{udpipe_tokenindex}
\alias{udpipe_tokenindex}
\title{Parse text using udpipe}
\usage{
udpipe_tokenindex(
  text,
  udpipe_model = "english-ewt",
  udpipe_model_path = getwd()
)
}
\arguments{
\item{text}{A character vector with text(s) to parse}

\item{udpipe_model}{Optionally, the name of a Universal Dependencies language model (e.g., "english-ewt", "dutch-alpino"), to use the udpipe package
(\code{\link[udpipe]{udpipe_annotate}}) for natural language processing. You can use \code{\link{show_udpipe_models}} to get
an overview of the available models. For more information about udpipe and performance benchmarks of the UD models, see the
GitHub page of the \href{https://github.com/bnosac/udpipe}{udpipe package}.}

\item{udpipe_model_path}{This path wil be used to look for the model, and if the model doesn't yet exist it will be downloaded to this location. Defaults to working directory}
}
\value{
A tokenIndex
}
\description{
Parse text using udpipe
}
