% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{unique_ids}
\alias{unique_ids}
\title{Make tokens ids globally unique (and update parents accordingly)}
\usage{
unique_ids(tokens)
}
\arguments{
\item{tokens}{a data frame of tokens}
}
\value{
the data frame with id and parent made globally unique
}
\description{
Make tokens ids globally unique (and update parents accordingly)
}

