% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_reshape.r
\name{unselect_nodes}
\alias{unselect_nodes}
\title{Undo select_nodes}
\usage{
unselect_nodes(.tokens)
}
\arguments{
\item{.tokens}{A tokenIndex in which nodes are selected with \link{select_nodes}.}
}
\description{
Not strictly required. Only available for elegance and minor memory efficiency
}
