% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dutch_alpino.r
\name{alpino_quote_queries}
\alias{alpino_quote_queries}
\title{Returns a list with the quote queries for ALPINO}
\usage{
alpino_quote_queries()
}
\value{
A list with rynstax queries, as created with \link{tquery}
}
\description{
Returns a list with the quote queries for ALPINO
}
