% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.r
\name{annotate}
\alias{annotate}
\title{Annotate a tokenlist based on rsyntax queries}
\usage{
annotate(tokens, queries, column, use = NULL, fill = T, block = NULL,
  check = T, with_tquery = F, show_fill = F)
}
\arguments{
\item{tokens}{A tokenIndex data.table, created with \link{as_tokenindex}, or any data.frame with the required columns (see \link{tokenindex_columns}).}

\item{queries}{A tquery or a list of queries, as created with \link{tquery}.}

\item{column}{The name of the column in which the annotations are added. The unique ids are added as [column]_id}

\item{fill}{If TRUE, the children for each id are added recursively (children of children etc.). If this leads to
duplicate ids (if an id in nodes is a child of another id in nodes), the most direct children are kept.
For example, if 1 -> 2 -> 3, and both 1 and 2 are in 'nodes', then 3 is only added as a child of 2.}

\item{block}{Optionally, a \link{subset} expression (e.g., quote == 'source') to block certain nodes from being assigned. For example, to prevent nodes to be extracted as clause subjects if they have already been extracted as quote sources.}

\item{check}{For testing queries. If TRUE, give a warning if there are duplicates in the data (in which case duplicates are deleted)}

\item{with_tquery}{For testing queries. If TRUE, add a column that shows the name of the specific tquery that was used. This only works if 'queries' is a named list.}
}
\description{
Apply queries to extract syntax patterns, and add the results as two columns to a tokenlist.
One column contains the ids for each hit. The other column contains the annotations.
Only nodes that are given a name in the tquery (using the 'save' parameter) will be added as annotation.
}
\details{
Note that while queries only find 1 node for each saved component of a pattern (e.g., quote queries have 1 node for "source" and 1 node for "quote"), 
all children of these nodes are also annotated (if fill is TRUE). If a child has multiple ancestors, only the most direct ancestors are used (see documentation for the fill argument).
}
