% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.r
\name{annotate_nodes}
\alias{annotate_nodes}
\title{Annotate a tokenlist based on rsyntaxNodes}
\usage{
annotate_nodes(tokens, nodes, column)
}
\arguments{
\item{tokens}{A tokenIndex data.table, or any data.frame coercible with \link{as_tokenindex}.}

\item{nodes}{A data.table, as created with \link{apply_queries}. Can be a list of multiple data.tables.}

\item{column}{The name of the column in which the annotations are added. The unique ids are added as [column]_id}
}
\description{
Use rsyntaxNodes, as created with \link{tquery} and \link{apply_queries}, to annotate a tokenlist.
Two columns will be added.
One column contains the ids for each hit. The other column contains the annotations.
Only nodes that are given a name in the tquery (using the 'label' parameter) will be added as annotation.
}
