% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.r
\name{annotate_qc}
\alias{annotate_qc}
\title{Annotate quotes and clauses}
\usage{
annotate_qc(tokens, quote_rules, clause_rules, check = F, with_rule = F)
}
\arguments{
\item{tokens}{A tokenIndex data.table, created with \link{as_tokenindex}, or any data.frame with the required columns (see \link{tokenindex_columns}).}

\item{quote_rules}{A list with rules for extracting quotes, as created with \link{rule}. For example, see \link{alpino_quote_rules}.}

\item{clause_rules}{A list with rules for extracting clauses, as created with \link{rule}. For example, see \link{alpino_clause_rules}.}

\item{check}{For testing rules. If TRUE, give a warning if there are duplicates in the data (in which case duplicates are deleted)}

\item{with_rule}{For testing rules. If TRUE, add a column that shows the name of the specific rule that was used. This only works if the rules are passed as a named list.}
}
\description{
A convenience function that calls the \link{annotate} function for a list of quote and clause rules.
}
\details{
This function assumes that sources (of quotes) cannot also be subjects (of clauses). Therefore, the quote rules are applied first,
and source nodes are blocked when applying the clause rules.
}
