% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clauses.R, R/clauses_alpino.R
\name{get_clauses}
\alias{get_clauses}
\title{Get the clauses for the given sentences}
\usage{
get_clauses(tokens, quotes = NULL)

get_clauses(tokens, quotes = NULL)
}
\arguments{
\item{tokens}{a data frame of tokens containing id, parent, relation, pos1 and lemma columns.}

\item{quotes}{optionally, the quotes in long format, i.e. a data frame with quote id, quote role, and token id}

\item{tokens}{a data frame of tokens containing id, parent, relation, pos1 and lemma columns.}

\item{quotes}{optionally, the quotes in long format, i.e. a data frame with quote id, quote role, and token id}
}
\value{
the clauses in long format, i.e. a data frame with clause id, clause role, and token id

the clauses in long format, i.e. a data frame with clause id, clause role, and token id
}
\description{
Runs the grammar rules on the tokens to get clauses (subject - predicate pairs).
If you give the quotes  (as returned by get.quotes), they will be excluded from being clauses as well.
Note: the token ids should be globally unique, see unique_ids. 
If the tokens contain a boolean 'attack' column, nouns with attack=T are considered for nominal actions

Runs the grammar rules on the tokens to get clauses (subject - predicate pairs).
If you give the quotes  (as returned by get.quotes), they will be excluded from being clauses as well.
Note: the token ids should be globally unique, see unique_ids. 
If the tokens contain a boolean 'attack' column, nouns with attack=T are considered for nominal actions
}

