% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_nodes.r
\name{get_long_ids}
\alias{get_long_ids}
\title{Get ids in various forms to extract token_ids}
\usage{
get_long_ids(..., select = NULL, with_fill = FALSE)
}
\arguments{
\item{...}{Either a data.table with the columns doc_id, sentence and token_id, or the output of \link{apply_queries}}

\item{select}{If not null, a character vector for selecting column names}

\item{with_fill}{If TRUE, include the ids of the fill nodes}
}
\value{
A data.table with the columns doc_id, sentence and token_id
}
\description{
Get ids in various forms to extract token_ids
}
