% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/english_corenlp.r
\name{get_quotes}
\alias{get_quotes}
\title{Get quotes from tokens parsed by coreNLP}
\usage{
get_quotes(tokens)
}
\arguments{
\item{tokens}{a token list data frame}

\item{block}{Optionally, .G_ID's to exclude from search. Can also be a data.table with nodes, as returned by
find_nodes or get_quotes_alpino}
}
\value{
a data.table with nodes (as .G_ID) for id, source and quote
}
\description{
Get quotes from tokens parsed by coreNLP
}
