% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsyntax.R
\name{graph_from_sentence}
\alias{graph_from_sentence}
\title{Create an igraph tree from a sentence}
\usage{
graph_from_sentence(tokens)
}
\arguments{
\item{tokens}{a data frame of tokens containing id, parent, and relation columns. If there are also quote and clause columns
(as output by annotate_tokens), the graph will be annotated with these relations.}
}
\value{
an igraph graph
}
\description{
Create an igraph tree from a data frame with token information, e.g, as donwloaded with amcat.get.tokens.
This assumes that id, parent, and relation are presente in the data frame. 
Any other columns will be available as vertex and edge attributes (edge attributes are taken from the child token).
If available, the 'word' column will be used to label the vertices.
The data frame should contain a single sentence with a single root node (with no parent).
}

