% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tquery.r
\name{NOT}
\alias{NOT}
\title{Use NOT search in tquery}
\usage{
NOT(...)
}
\arguments{
\item{...}{name-value pairs for lookup terms. see ?query.}
}
\value{
A list, to be used as input to \link{tquery}
}
\description{
Use NOT search in tquery
}
\examples{
tquery(NOT(POS='Noun'))   
}
