% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tquery.r
\name{OR}
\alias{OR}
\title{Use OR search in tquery}
\usage{
OR(...)
}
\arguments{
\item{...}{name-value pairs for lookup terms. see ?query.}
}
\value{
A list, to be used as input to \link{tquery}
}
\description{
Use OR search in tquery
}
\examples{
tquery(OR(lemma = 'walk', POS='Noun'))
}
