% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.r
\name{plot_tree}
\alias{plot_tree}
\title{Create an igraph tree from a sentence}
\usage{
plot_tree(tokens, root_i = 1, label_vars = c("token"),
  quote_var = "quote", clause_var = "clause")
}
\arguments{
\item{tokens}{A tokenIndex data.table, created with \link{as_tokenindex}, or any data.frame with the required columns (see \link{tokenindex_columns}).}

\item{root_i}{By default, plot_tree uses the first root (i.e. token without parent) in the data. Change root_i to use the second (2), third (3), etc.}

\item{label_vars}{A character vector, specifying the columns that will be used to label the nodes}

\item{quote_var}{The name of the column with quote annotations (will be ignored if column is not available)}

\item{clause_var}{The name of the column with quote annotations (will be ignored if column is not available)}
}
\value{
an igraph graph
}
\description{
Create an igraph tree from a token_index (\link{as_tokenindex}) or a data.frame that can be coerced to a tokenindex.
}
