% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.r
\name{plot_tree}
\alias{plot_tree}
\title{Create an igraph tree from a sentence}
\usage{
plot_tree(tokens, sentence_i = 1, doc_id = NULL, sentence = NULL,
  quote_var = "quote", clause_var = "clause", coref = "coref_id",
  label_size = 0.9, node_size = 30, edge_label_size = 0.8)
}
\arguments{
\item{tokens}{A tokenIndex data.table, created with \link{as_tokenindex}, or any data.frame with the required columns (see \link{tokenindex_columns}).}

\item{sentence_i}{By default, plot_tree uses the first sentence (sentence_i = 1) in the data. sentence_i can be changed to select other sentences by position (the i-th unique sentence in the data). Note that sentence_i does not refer to the values in the sentence column (for this use the sentence argument together with doc_id)}

\item{doc_id}{Optionally, the document id can be specified. If so, sentence_i refers to the i-th sentence within the given document.}

\item{sentence}{Optionally, the sentence id can be specified (note that sentence_i refers to the position). If sentence is given, doc_id has to be given as well.}

\item{quote_var}{The name of the column with quote annotations. Will be ignored if column is not available.}

\item{clause_var}{The name of the column with quote annotations. Will be ignored if column is not available.}

\item{label_var}{The name of the column with the token label (word or lemma). Will be ignored if column is not available.}

\item{pos_var}{The name of the column with the part-of-speech tag. Will be ignored if column is not available.}
}
\value{
an igraph graph
}
\description{
Create an igraph tree from a token_index (\link{as_tokenindex}) or a data.frame that can be coerced to a tokenindex.
}
