% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tquery.r
\name{print.tQuery}
\alias{print.tQuery}
\title{S3 print for tQuery class}
\usage{
\method{print}{tQuery}(x, ...)
}
\arguments{
\item{x}{a tQuery}

\item{...}{not used}
}
\description{
S3 print for tQuery class
}
\examples{
q = tquery(select = lemma \%in\% .VIND_VERBS, 
                   children(save = 'source', p_rel=.SUBJECT_REL),
                   children(p_rel='vc', select = POS \%in\% c('C', 'comp'),
                            children(save='quote', p_rel=.SUBJECT_BODY)))
q 
}
