% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.r
\name{print.tQuery}
\alias{print.tQuery}
\title{S3 print for tQuery class}
\usage{
\method{print}{tQuery}(x, ...)
}
\arguments{
\item{x}{a tQuery}

\item{...}{not used}
}
\description{
S3 print for tQuery class
}
\examples{
q = tquery(label='quote',
           children(relation='nmod:according_to', label='source',
                    children(label='verb')))
q 
}
