% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_index.r
\name{tokenindex_columns}
\alias{tokenindex_columns}
\title{Specify the column names used in rsyntax.}
\usage{
tokenindex_columns(doc_id = "doc_id", sentence = "sentence",
  token_id = "token_id", parent = "parent")
}
\arguments{
\item{doc_id}{document id}

\item{token_id}{position or id of the token (document-sentence-token must be unique)}

\item{parent}{the token id of the parent}

\item{sentence_id}{position or id of sentence (in document)}
}
\description{
Certain columns are required for applying (or writing) rsyntax queries. With the default column names, these are:
document id ("doc_id"), position or id of the token ("token_id"), 
and the token id of the parent ("parent").
}
\details{
With this function, you can specify aliases instead of these default names.
}
