% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_text.r
\name{udpipe_tokenindex}
\alias{udpipe_tokenindex}
\title{Parse text using udpipe}
\usage{
udpipe_tokenindex(text, udpipe_model = "english",
  udpipe_model_path = getOption("corpustools_resources", NULL))
}
\arguments{
\item{text}{A character vector with text(s) to parse}

\item{udpipe_model}{The name of a udpipe language model (e.g., "english", "dutch", "german"), to use the udpipe package to perform natural language processing. On first use, the model will be downloaded to the location specified in the udpipe_model_path argument.}

\item{udpipe_model_path}{If udpipe_model is used, this path wil be used to look for the model, and if the model doesn't yet exist it will be downloaded to this location. If no path is given, the directory in which the \link{corpustools} package was installed will be used.}
}
\description{
Parse text using udpipe
}
