% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_reshape.r
\name{unselect_nodes}
\alias{unselect_nodes}
\title{Undo select_nodes}
\usage{
unselect_nodes(.tokens)
}
\arguments{
\item{.tokens}{A tokenIndex in which nodes are selected with \link{select_nodes}.}
}
\value{
A tokenIndex (without a .nodes attribute)
}
\description{
Not strictly required. Only available for elegance and minor memory efficiency
}
\examples{
tokens = tokens_spacy[tokens_spacy$doc_id == 'text4',]

tq = tquery(relation = "relcl", label = "relative_clause")
tokens = select_nodes(tokens, tq) 
selected_nodes(tokens)

tokens = unselect_nodes(tokens)

is.null(attr(tokens, '.nodes'))
}
